Issue #4437: Modified ImportControlCheckTest.java and moved its input files to the importcontrol subdirectory
