+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/imports/importcontrol/InputImportControl.java
@@ -1,4 +1,4 @@
-package com.puppycrawl.tools.checkstyle.checks.imports;
+package com.puppycrawl.tools.checkstyle.checks.imports.importcontrol;
 
 import java.awt.Image;
 import javax.swing.border.*;
