+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/imports/importcontrol/InputImportControlBlacklist.xml
@@ -3,9 +3,9 @@
         "-//Puppy Crawl//DTD Import Control 1.2//EN"
         "http://checkstyle.sourceforge.net/dtds/import_control_1_2.dtd">
 
-<import-control pkg="com.puppycrawl.tools.checkstyle.checks">
+<import-control pkg="com.puppycrawl.tools.checkstyle.checks.imports">
   <allow pkg=".*" regex="true"/>
-  <subpackage name="imports">
+  <subpackage name="importcontrol">
     <disallow class="java.util.Date"/>
     <disallow pkg="java.util.stream"/>
   </subpackage>
