+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/imports/importcontrol/InputImportControlOne.xml
@@ -3,12 +3,11 @@
     "-//Puppy Crawl//DTD Import Control 1.0//EN"
     "http://checkstyle.sourceforge.net/dtds/import_control_1_0.dtd">
 
-<import-control pkg="com.puppycrawl.tools.checkstyle.checks">
+<import-control pkg="com.puppycrawl.tools.checkstyle.checks.imports">
   <allow class="java.awt.Image"/>
+  <allow class="java.awt.Button.ABORT"/>
   <allow class="java.io.File" local-only="true"/>
-  <subpackage name="imports">
-    <disallow class="java.awt.Image" local-only="true"/>
-    <allow pkg="javax.swing" exact-match="true"/>
-    <allow pkg="java.io" exact-match="true" local-only="true"/>
+  <subpackage name="importcontrol">
+    <allow pkg="javax.swing"/>
   </subpackage>
 </import-control>
