+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/imports/importcontrol/InputImportControlOneRegExp.xml
@@ -1,0 +1,13 @@
+<?xml version="1.0"?>
+<!DOCTYPE import-control PUBLIC
+    "-//Puppy Crawl//DTD Import Control 1.2//EN"
+    "http://checkstyle.sourceforge.net/dtds/import_control_1_2.dtd">
+
+<import-control pkg="com.puppycrawl.tools.checkstyle.checks.imports">
+  <allow class="java\.awt.Image" regex="true"/>
+  <allow class="java..*.Button.ABORT" regex="true"/>
+  <allow class="java.(io|lui).File" local-only="true" regex="true"/>
+  <subpackage name="importcontrol">
+    <allow pkg="javax.swing" regex="true"/>
+  </subpackage>
+</import-control>
