+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/imports/importcontrol/InputImportControlPkgRegExpInChild.xml
@@ -3,11 +3,11 @@
     "-//Puppy Crawl//DTD Import Control 1.2//EN"
     "http://checkstyle.sourceforge.net/dtds/import_control_1_2.dtd">
 
-<import-control pkg="com.puppycrawl.tools.checkstyle.checks">
+<import-control pkg="com.puppycrawl.tools.checkstyle.checks.imports">
   <allow class="java.awt.Image"/>
   <allow class="java.awt.Button.ABORT"/>
   <allow class="java.io.File" local-only="true"/>
-  <subpackage name="imp.rts" regex="true">
+  <subpackage name="imp.rtcontrol" regex="true">
     <allow pkg="javax.swing"/>
   </subpackage>
 </import-control>
