+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/imports/importcontrol/InputImportControlStrategyOnMismatchOne.xml
@@ -1,12 +1,12 @@
 <?xml version="1.0"?>
 <!DOCTYPE import-control PUBLIC
-    "-//Puppy Crawl//DTD Import Control 1.2//EN"
-    "http://checkstyle.sourceforge.net/dtds/import_control_1_2.dtd">
+    "-//Puppy Crawl//DTD Import Control 1.3//EN"
+    "http://checkstyle.sourceforge.net/dtds/import_control_1_3.dtd">
 
-<import-control pkg="com.puppycrawl.tools.checkstyle.checks">
+<import-control pkg="com.puppycrawl.tools.checkstyle.checks.imports" strategyOnMismatch="disallowed">
   <allow class="java\.awt\.Image" regex="true"/>
   <allow class="java\..*\.File" local-only="true" regex="true"/>
-  <subpackage name="imports">
+  <subpackage name="importcontrol" strategyOnMismatch="delegateToParent">
     <disallow class="java\..*\.Image" local-only="true" regex="true"/>
     <allow pkg="javax\.swing" exact-match="true" regex="true"/>
     <allow pkg="java\.io" exact-match="true" local-only="true" regex="true"/>
