+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/imports/importcontrol/InputImportControlStrategyOnMismatchThree.xml
@@ -3,8 +3,8 @@
     "-//Puppy Crawl//DTD Import Control 1.3//EN"
     "http://checkstyle.sourceforge.net/dtds/import_control_1_3.dtd">
 
-<import-control pkg="com.puppycrawl.tools.checkstyle.checks" strategyOnMismatch="allowed">
-  <subpackage name="imports" strategyOnMismatch="delegateToParent">
+<import-control pkg="com.puppycrawl.tools.checkstyle.checks.imports" strategyOnMismatch="allowed">
+  <subpackage name="importcontrol" strategyOnMismatch="delegateToParent">
     <disallow class="java\..*\.Image" local-only="true" regex="true"/>
   </subpackage>
 </import-control>
