+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/imports/importcontrol/InputImportControlStrategyOnMismatchTwo.xml
@@ -3,12 +3,12 @@
     "-//Puppy Crawl//DTD Import Control 1.3//EN"
     "http://checkstyle.sourceforge.net/dtds/import_control_1_3.dtd">
 
-<import-control pkg="com.puppycrawl.tools.checkstyle.checks" strategyOnMismatch="disallowed">
+<import-control pkg="com.puppycrawl.tools.checkstyle.checks.imports" strategyOnMismatch="allowed">
   <allow class="java\.awt\.Image" regex="true"/>
   <allow class="java\..*\.File" local-only="true" regex="true"/>
-  <subpackage name="imports" strategyOnMismatch="delegateToParent">
+  <subpackage name="importcontrol" strategyOnMismatch="disallowed">
     <disallow class="java\..*\.Image" local-only="true" regex="true"/>
-    <allow pkg="javax\.swing" exact-match="true" regex="true"/>
+    <allow pkg="javax\.swing" regex="true"/>
     <allow pkg="java\.io" exact-match="true" local-only="true" regex="true"/>
   </subpackage>
 </import-control>
