+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/imports/importcontrol/InputImportControlTwoRegExp.xml
@@ -1,14 +1,14 @@
 <?xml version="1.0"?>
 <!DOCTYPE import-control PUBLIC
-    "-//Puppy Crawl//DTD Import Control 1.3//EN"
-    "http://checkstyle.sourceforge.net/dtds/import_control_1_3.dtd">
+    "-//Puppy Crawl//DTD Import Control 1.2//EN"
+    "http://checkstyle.sourceforge.net/dtds/import_control_1_2.dtd">
 
-<import-control pkg="com.puppycrawl.tools.checkstyle.checks" strategyOnMismatch="allowed">
+<import-control pkg="com.puppycrawl.tools.checkstyle.checks.imports">
   <allow class="java\.awt\.Image" regex="true"/>
   <allow class="java\..*\.File" local-only="true" regex="true"/>
-  <subpackage name="imports" strategyOnMismatch="disallowed">
+  <subpackage name="importcontrol">
     <disallow class="java\..*\.Image" local-only="true" regex="true"/>
-    <allow pkg="javax\.swing" regex="true"/>
+    <allow pkg="javax\.swing" exact-match="true" regex="true"/>
     <allow pkg="java\.io" exact-match="true" local-only="true" regex="true"/>
   </subpackage>
 </import-control>
